Implement algorithm to compare different hand values
Every hand value has a ranking represented by a simple int with a
straight flush being the highest and high card the lowest.

It first compares the hand by this value.

If it has the same hand value type, e.g. both hands holding a three of
a kind, it then calculates a specific hand value - also a simple number
that represents the rank of this specific hand value.

Calculation is specific to the hand value and differs slightly between
them. For example: Three of a kinds are ranked by the value of their triplets
, then by the rank of the high value kicker and finally by the rank of
the low value kicker. In this order every value is multiplied by a
decreasing factor that represents the priority.

those factors are squares of 13 because we have 13 different card
values. If we'd choose squares of 10 a lower ranked card value could
have a higher numeric value, than a higher ranked card value.